fix(AppNodeSchema): Fix type of Git metadata
`PendingNodeData::defaultRev` was defined as `GitMetadata | undefined`
while the `AppNodeSchema` defined it as `nullable`.
When creating a node without setting a Git revision this leads to
a form validation error.

This change sets `PendingNodeData::defaultRev` as `GitMetadata |
undefined` to align all usages.
