+++ frontend/src/store.ts
@@ -302,7 +302,7 @@ export const useStore = create<AppState>()(
           eventScreenPosition: { x: clientX, y: clientY },
           type: nodeType,
           fromNodeId: fromNode.id,
-          defaultRev: undefined,
+          defaultRev: null,
         });
       },
       setEdgeType: (newType) => {
